Hamiltonian Cycle instead of ANN
Implemented a different form of AI that creates a hamiltonian cycle on the playfield that it follows instead. THis ensures that it will never run into it's tail or a wall.

Also implemented a way for it to jump forward on the path if certain conditions are met. As well as a point system that discourages it from jumping forward if the tail end is too close.